+++ src/accessories/temperature.js
@@ -11,7 +11,7 @@ class TemperatureAccessory {
     try {
       const pkg = require('../../package.json');
       const info = this.accessory.getService(this.api.hap.Service.AccessoryInformation);
-      info.setCharacteristic(this.api.hap.Characteristic.Manufacturer, 'Homebridge WeatherXM');
+      info.setCharacteristic(this.api.hap.Characteristic.Manufacturer, 'Homebridge TC WeatherXM');
       info.setCharacteristic(this.api.hap.Characteristic.Model, 'WeatherXM Sensor');
       info.setCharacteristic(this.api.hap.Characteristic.SerialNumber, `${this.platform.client.stationId || 'unknown'}-temperature`);
       info.setCharacteristic(this.api.hap.Characteristic.FirmwareRevision, pkg.version || '0.0.0');